Did some code cleanup
