+++ engine/Components/ErrorHandler.py
@@ -1,0 +1,1 @@
+#todo handles with errors from
