+++ engine/Components/Sanitizer.py
@@ -1,0 +1,5 @@
+#todo changes statement to something that we can use
+#todo change to something that the system can use
+#todo add brackets
+
+# for example: change & to ^, | to v
