+++ engine/Components/Verifier.py
@@ -1,0 +1,2 @@
+#todo check connectives
+#todo suggest list of connectives
