+++ engine/Inputs/UserInput.py
@@ -1,19 +1,25 @@
+from engine.Components.ParseFunction import parse_statements
 from engine.Inputs.input_statements import statement as instate
 from engine.Components.AnalizeFunction import analize_statement
 
 
+# returns an integer of the number of variables,
+# an array and string of the variables respectively,
+# a statement that is provided and an array with the individual elements
 def user_input():
     statement = instate
 
-    print("In user_input():")
+    # print("In user_input():")
     # statement = input("Please enter the statement(s): ")
-    print("Please enter the statement(s): P ^ Q, R v (P -> Q), R ^ (Q v P) <-> (R ^ Q) v (R ^ P)")
-    # statement = "P ^ Q, R v (P -> Q), R ^ (Q v P) <-> (R ^ Q) v (R ^ P) ^ T"
+    print("Please enter the statement(s): " + statement)
 
-
-    # analize statement to determine the number of variables
-    # todo move analize_statement() to separate file
+    # analyze statement to determine the number of variables,
+    # and split into an array of elements
     number_of_variables, variables_arr, variables_str, elements = analize_statement(statement)
 
-    print("Out user_input():")
-    return number_of_variables, variables_arr, variables_str, statement, elements
+    # Combine variables and connectives
+    # print("\nCalling parse_statements():")
+    new_elements = parse_statements(statement, elements, variables_arr)
+
+    # print("Out user_input():")
+    return number_of_variables, variables_arr, variables_str, statement, new_elements
