+++ engine/Main.py
@@ -3,8 +3,7 @@ from engine.Inputs.UserInput import *
 # main()
 print("***********************************************************")
 print("In main():")
-# User Input, Calls user_input() function in LogicCalculator.py
-# todo Move user_input() to new file.
+
 # creates and initializes variables
 num_of_variables, variables_arr, variables_str, statement, elements = user_input()
